Add files via upload
This is the merged Prototype of the application. In this version the introductory code and issues menu's have all been linked. 
Code-wise, formatting issues have been resolved as best as possible (i.e. times where people had attempted lists or bold/italicized fonts have been reformatted to work as per Markdown). A major piece that we will need to work out for the S2 is the 'return to menu' option. We have included it in our writing here however in code the 'return to menu' buttons are all 'restart' buttons which take the user back to the first screen of the entire program. We will need to resolve this however my attempts to fix it with an event etc have been unsuccessful.